requirements.txt
Cython
ipykernel
matplotlib
opencv-python-headless
pytorch-lightning
scipy
torchvision
wandb
shapely